requirements.txt
better-profanity==0.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==2.2.2
gunicorn==20.0.4
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
nltk==3.8.1
PyPDF2==3.0.1
pyspellchecker==0.8.1
regex==2024.7.24
requests==2.32.3
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==2.2.2
zipp==3.19.2
